+++ ChangeLog.md
@@ -1,6 +1,7 @@
 #2.2.11
   * Add help system and initial help file
   * Make absolute damage visible
+  * Add 'average' roll for blueprints
 
 #2.2.10
   * Fix detailed export of module reinforcement packages